copied registration survey over to registration.tsx, and created the signin page in signin.tsx
