+++ frontend/src/pages/Registration/RegistrationPage.module.css
@@ -1,3 +1,37 @@
+.title {
+    font-size: 34px;
+    font-family: 'Roboto', sans-serif;
+    text-align: center;
+    font-weight: 700;
+}
+
+.content {
+    font-family: 'Roboto', sans-serif;
+    text-align: center;
+}
+
+.confirmPasswordWarning {
+    font-family: 'Roboto', sans-serif;
+    color: red;
+    font-size: 12px;
+    margin-left: 3px;
+    margin-top: 5px;
+}
+
 .container {
-    
+    max-width: 100vw;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin: 25px;
+}
+
+.containerWrapper {
+    max-width: 100vw;
+    width: 50%;
+    display: flex;
+    flex-direction: column;
+    gap: 50px;
 }
